+++ stko/utilities/utilities.py
@@ -1730,3 +1730,21 @@ def get_atom_distance(position_matrix, atom1_id, atom2_id):
     )
 
     return float(distance)
+
+
+def get_long_bond_ids(mol):
+    """
+    Return tuple of long bond ids in a ConstructedMolecule.
+
+    """
+
+    long_bond_ids = []
+    for bond_infos in mol.get_bond_infos():
+        if bond_infos.get_building_block() is None:
+            ids = (
+                bond_infos.get_bond().get_atom1().get_id(),
+                bond_infos.get_bond().get_atom2().get_id(),
+            )
+            long_bond_ids.append(ids)
+
+    return tuple(long_bond_ids)
